"Cleaned the code a bit. Small bug: can't open devtools with custom titlebar..."
